refactor: extract useTranscriptionSummary for shared summary state/logic
TranscriptionPanel (dashboard, Card/Button chrome) and
recordings/TranscriptionSection (Panel/MetalButton chrome) had the
same summary fetch + invalidation + handlers duplicated across both
files. The visual chrome is deliberately different across the two
surfaces -- merging the render would flatten the design difference,
so only the logic dedupes cleanly.

Extract to useTranscriptionSummary(recordingId, transcriptionText)
returning flat state + handlers. Both components keep their own
JSX and the dashboard's shadcn look stays distinct from the
recording page's metal look.

Drops ~80 duplicated lines across the pair. No behavior change.
